add check for directory

+++ extractor.py
@@ -31,7 +31,12 @@ class Extractor:
         target_file_list = []
         extraction_id: str = str(uuid4())
         
-        os.mkdir(f'tmp/{extraction_id}')
+        tmp_dir = f'tmp/{extraction_id}'
+        result_dir = f'result/{extraction_id}'
+        inkscape_log_dir = f'inkscape_log/{extraction_id}'
+        os.mkdir(tmp_dir)
+        os.mkdir(inkscape_log_dir)
+
         try:
             with open(file_path, "rb") as f:
                 input_pdf = PdfFileReader(f)
@@ -44,7 +49,7 @@ class Extractor:
 
                     # if it already exists, then it's time to clear the directory
                     
-                    with open(f'tmp/{extraction_id}/{i}.ai', 'wb') as temp_f:
+                    with open(f'{tmp_dir}/{i}.ai', 'wb') as temp_f:
                         output.write(temp_f)
 
                     inkscape_actions.append(self.INKSCAPE_EXPORT_ACTIONS.format(
@@ -58,7 +63,7 @@ class Extractor:
 
             inkscape_actions_str = ";".join(inkscape_actions)
             if inkscape_log:
-                with open(f"inkscape_log/{extraction_id}.out.txt", "wb") as f_out, open(f"inkscape_log/{extraction_id}.err.txt", "wb") as f_err: 
+                with open(f"{inkscape_log_dir}/out.txt", "wb") as f_out, open(f"{inkscape_log_dir}/err.txt", "wb") as f_err: 
                     proc = Popen(shlex.split(f'{self.INKSCAPE_PATH} --actions="{inkscape_actions_str}"'), stdout=f_out, stderr=f_err)
             else:
                 proc = Popen(shlex.split(f'{self.INKSCAPE_PATH} --actions="{inkscape_actions_str}"'), stdout=DEVNULL, stderr=DEVNULL)
@@ -69,11 +74,15 @@ class Extractor:
 
                 if cleanup:
                     for file_path in target_file_list:
-                        os.remove(file_path)
-                    shutil.rmtree(f'tmp/{extraction_id}')
+                        if os.path.exists(file_path):
+                            os.remove(file_path)
+                    if os.path.exists(tmp_dir):
+                        shutil.rmtree(tmp_dir)
             return extraction_id
         except Exception as e:
             if cleanup:
-                shutil.rmtree(f'tmp/{extraction_id}')
-                shutil.rmtree(f'result/{extraction_id}')
+                if os.path.exists(tmp_dir):
+                    shutil.rmtree(tmp_dir)
+                if os.path.exists(result_dir):
+                    shutil.rmtree(result_dir)
             raise e
